type changes

+++ src/cf3globals.c
@@ -38,16 +38,16 @@
 /* flags                                                                     */
 /*****************************************************************************/
 
-short SHOWREPORTS = false;
+int SHOWREPORTS = false;
 
 /*****************************************************************************/
 /* operational state                                                         */
 /*****************************************************************************/
 
-short VERBOSE = false;
-short INFORM = false;
-short PARSING = false;
-short CFPARANOID = false;
+int VERBOSE = false;
+int INFORM = false;
+int PARSING = false;
+int CFPARANOID = false;
 int REQUIRE_COMMENTS = CF_UNDEFINED;
 
 struct utsname VSYSNAME;
@@ -55,7 +55,7 @@ struct utsname VSYSNAME;
 FILE *FREPORT_HTML = NULL;
 FILE *FREPORT_TXT = NULL;
 FILE *FKNOW = NULL;
-short XML = false;
+int XML = false;
 struct FnCallStatus FNCALL_STATUS;
 
 int CFA_MAXTHREADS = 10;
@@ -66,7 +66,7 @@ int AM_BACKGROUND_PROCESS = false;
 char *THIS_BUNDLE = NULL;
 char THIS_AGENT[CF_MAXVARSIZE];
 enum cfagenttype THIS_AGENT_TYPE;
-short INSTALL_SKIP = false;
+int INSTALL_SKIP = false;
 int FACILITY;
 time_t PROMISETIME;
 
@@ -204,13 +204,13 @@ char *CF_AGENTTYPES[] = /* see enum cfagenttype */
 
 double FORGETRATE = 0.7;
 
-short IGNORELOCK = false;
-short DONTDO = false;
-short DEBUG = false;
-short D1 = false;
-short D2 = false;
-short AUDIT = false;
-short LOGGING = false;
+int IGNORELOCK = false;
+int DONTDO = false;
+int DEBUG = false;
+int D1 = false;
+int D2 = false;
+int AUDIT = false;
+int LOGGING = false;
 
 char  VFQNAME[CF_MAXVARSIZE];
 char  VUQNAME[CF_MAXVARSIZE];
@@ -361,18 +361,18 @@ int RPCTIMEOUT = 60;          /* seconds */
 pid_t ALARM_PID = -1;
 int SENSIBLEFILECOUNT = 2;
 int SENSIBLEFSSIZE = 1000;
-short SKIPIDENTIFY = false;
-short ALL_SINGLECOPY = false;
-short FULLENCRYPT = false;
+int SKIPIDENTIFY = false;
+int ALL_SINGLECOPY = false;
+int FULLENCRYPT = false;
 int EDITFILESIZE = 10000;
-short NOHARDCLASSES=false;
+int NOHARDCLASSES=false;
 int VIFELAPSED = 1;
 int VEXPIREAFTER = 120;
-short UNDERSCORE_CLASSES=false;
+int UNDERSCORE_CLASSES=false;
 int CHECKSUMUPDATES = false;
 char BINDINTERFACE[CF_BUFSIZE];
-short MINUSF = false;
-short EXCLAIM = true;
+int MINUSF = false;
+int EXCLAIM = true;
 
 mode_t DEFAULTMODE = (mode_t) 0755;
 
@@ -383,7 +383,7 @@ struct Item *VDEFAULTROUTE=NULL;
 struct Item *VSETUIDLIST = NULL;
 struct Item *SUSPICIOUSLIST = NULL;
 enum classes VSYSTEMHARDCLASS = unused1;
-short NONALPHAFILES = false;
+int NONALPHAFILES = false;
 struct Item *EXTENSIONLIST = NULL;
 struct Item *SPOOLDIRLIST = NULL;
 struct Item *NONATTACKERLIST = NULL;
